Extract database interaction logic from getServerSideProps in pages/index.ts file, and create getReducedProjects util function to retrieve reduced projects for the showcase section.
